vp: core: rv64: csr: fixed csr_pmpaddr getter
Bug-report and fix by Thomas Arndt
https://github.com/ics-jku/riscv-vp-plusplus/issues/14
"
struct csr_pmpaddr {
	union reg {
		uint64_t val = 0;
		struct fields {
			uint64_t address : 54;
			unsigned zero : 10;
		} fields;
	} reg;

	unsigned get_address() {
		return reg.fields.address << 2;
	}
};

fixed - the get_address() better return a 'uint64_t' instead of 'unsigned', otherwise the upper 32bit are truncated.

struct csr_pmpaddr {
	union reg {
		uint64_t val = 0;
		struct fields {
			uint64_t address : 54;
			unsigned zero : 10;
		} fields;
	} reg;

	uint64_t get_address() {
		return reg.fields.address << 2;
	}
};
"

+++ vp/src/core/rv64/csr.h
@@ -241,7 +241,7 @@ struct csr_pmpaddr {
 		} fields;
 	} reg;
 
-	unsigned get_address() {
+	uint64_t get_address() {
 		return reg.fields.address << 2;
 	}
 };
